requirements-txt.txt
# Core dependencies with compatible versions
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
numpy<2.0.0

# Chemical database APIs
requests==2.31.0

# Development tools
pytest==7.4.3
black==23.11.0

# Future additions (uncomment when ready)
# fastapi==0.104.1
# uvicorn==0.24.0
# openai==1.3.7